### 2025-02-26 #690 @kelliejarcher
hdcuremodels
Submitting Author Name: Kellie J. Archer
Submitting Author Github Handle: @kelliejarcher
Repository:  https://github.com/kelliejarcher/hdcuremodels
Submission type: Pre-submission

---

```
Package: hdcuremodels
Title: Penalized Mixture Cure Models for High-Dimensional Data
Version: 0.0.2
    c(person("Han", "Fu", role = "aut"), person(c("Kellie J."), "Archer", email=
    "archer.43@osu.edu", role = c("aut","cre"), comment = c(ORCID="0000-0003-1555-5781")))
Description: Provides functions for fitting various penalized parametric and semi-parametric mixture cure models with different penalty functions, testing for a significant cure fraction, and testing for sufficient follow-up as described in Fu et al (2022)<doi:10.1002/sim.9513> and Archer et al (2024)<doi:10.1186/s13045-024-01553-6>. False discovery rate controlled variable selection is provided using model-X knock-offs.
Imports: doParallel,
         flexsurv,
         flexsurvcure,
         foreach,
         ggpubr,
         glmnet,
         knockoff,
         mvnfast,
         parallel,
         plyr,
         survival
Roxygen: list(markdown = TRUE, roclets = c ("namespace", "rd", "srr::srr_stats_roclet"))
    roxygen2
```

	- [X ] Regression and Supervised Learning

Yes, I tried to add notations via the srr package. This is all very new so advice would be most welcome.

Analysts interested in modeling a time-to-event outcome when a subset of patients experience long-term survival or cure. This package permits fitting penalized mixture cure models so the functions can handle modeling the time-to-event outcome when the covariate/predictor space is high-dimensional.

None of the existing packages that fit mixture cure models (MCMs) are capable of handling high-dimensional datasets. Only `penPHcure` includes a LASSO penalty to perform variable selection for scenarios when the sample size exceeds the number of predictors. Other R packages that can be used for fitting MCMs include:
* `cuRe` (Jakobsen, 2023) can be used to fit parametric MCMs on a relative survival scale;
* `CureDepCens` (Schneider and Grandemagne dos Santos, 2023) can be used to fit piecewise exponential or Weibull model with dependent censoring;
* `curephEM` (Hou and Ren, 2024) can be used to fit a MCM where the latency is modeled using a Cox PH model;
* `flexsurvcure` (Amdahl, 2022) can be used to fit parametric mixture and non-mixture cure models;
* `geecure` (Niu and Peng, 2018) can be used to fit marginal MCM for clustered survival data;
* `GORCure` (Zhou et al, 2017) can be used to fit generalized odds rate MCM with interval censored data;
* `mixcure` (Peng, 2020) can be used to fit non-parametric, parametric, and semiparametric MCMs;
* `npcure` (López-de-Ullibarri and López-Cheda, 2020) can be used to non-parametrically estimate incidence and latency;
* `npcurePK` (Safari et al, 2023) can be used to non-parametrically estimate incidence and latency when cure is partially observed;
* `penPHcure`(Beretta and Heuchenne, 2019) can be used to fit semi-parametric PH MCMs with time-varying covariates; and
* `smcure` (Cai et al 2022) can be used to fit semi-parametric (PH and AFT) MCMs.

I am very inexperienced using GitHub and have not used it for collaboration before. I did post a version of this package on CRAN last June and more recently learned about ROpenSci so the github version is my initial attempt to adhere to your standards. I have not submitted a peer-reviewed manuscript yet as I would prefer to have an ROpenSci review first. Also, when I ran pkgcheck and tried to look at the summary I received a message that read, "Error: No GNU global installation found." and was unclear how to proceed.
