### 2025-02-27 #690 @mpadge
Hi @kelliejarcher, and thank you for your pre-submission. No worries about inexperience - rOpenSci strives to be as welcoming and inclusive as possible, and to always be answer any questions you might have, and to help you along the way. As this is a pre-submission inquiry, feel free to ask questions here, or alternatively open issues in your own repository, cross-link them here by pasting the url for this issue in the comment, and ping me there. (Note that we try to keep the full submissions as "clean" as possible to help focus on reviews, while pre-submissions are the place for more general questions and dicussions.)

Specific responses to your questions:

- The "Error: No GNU global installation found" is because {pkgcheck}, and the {pkgstats} package it uses to analyses packages, require a couple of system libraries, including "GNU global". If you're on a Linux-based or MacOS, installation is simply, generally by using standard package manager (`apt-get`, `homebrew`, or whatever), to `install global`. If you're on Windows, it's tricker, but you could start with the links in [the {pkgstats} installation vignette](https://docs.ropensci.org/pkgstats/articles/installation.html).
- The {srr} question sounds more general, and is maybe best moved to a specific issue within your repo? If you ping me there, I'll happily help further.

More generally, your package definitely looks like a good fit for statistical software review, and definitely within the category you've already indicated. Looking forward to working towards a full submission!
