### 2025-02-27 #690 @kelliejarcher
Hi Mark,

I used homebrew to install global and it is in my search path, as verified by

global --version

global (GNU Global) 6.6.14

Powered by Berkeley DB 1.85 and SQLite3 3.49.1.

Copyright (c) 1996-2024 Tama Communications Corporation

License GPLv3+: GNU GPL version 3 or later http://www.gnu.org/licenses/gpl.html

This is free software; you are free to change and redistribute it.

There is NO WARRANTY, to the extent permitted by law.

I also restarted my machine but am still getting
> x <- pkgcheck (mydir)
Error: No GNU global installation found.

I also have ctags installed:

ctags --version

Universal Ctags 6.1.0, Copyright (C) 2015-2023 Universal Ctags Team

Universal Ctags is derived from Exuberant Ctags.

Exuberant Ctags 5.8, Copyright (C) 1996-2009 Darren Hiebert

  Compiled: Feb 17 2025, 07:54:50

  URL: https://ctags.io/

  Output version: 0.0

  Optional compiled features: +wildcards, +regex, +gnulib_fnmatch, +gnulib_regex, +iconv, +option-directory, +xpath, +json, +interactive, +yaml, +case-insensitive-filenames, +packcc, +optscript, +pcre2

But in R when I run ctags_test() I get
Error: No GNU global installation found.

Do you know why this doesn’t work?

Also, would you please provide more specific guidance about what you mean by “or alternatively open issues in your own repository, cross-link them here by pasting the url for this issue in the comment, and ping me there.”? Does that mean that under the repository, I click on <Issues>, click on <New issue> and then type whatever text and then send a link? If so, do you want individual issues entered separately or can you have one long laundry list of issues? Sorry for all the questions.

Best regards,
Kellie

From: mark padgham ***@***.***>
Date: Thursday, February 27, 2025 at 6:08 AM
To: ropensci/software-review ***@***.***>
Cc: Archer, Kellie ***@***.***>, Mention ***@***.***>
Subject: Re: [ropensci/software-review] hdcuremodels (Issue #690)
Hi @kelliejarcher, and thank you for your pre-submission. No worries about inexperience - rOpenSci strives to be as welcoming and inclusive as possible, and to always be answer any questions you might have, and to help you along the way. As

Hi @kelliejarcher<https://urldefense.com/v3/__https:/github.com/kelliejarcher__;!!KGKeukY!1cCR1ET6HXFR0W8HksqghUebuvDfxPa2AmphB0DMkwS6B8XR4aTFPY3kwAa-MmzbsMJbCBpvou9AHHRYSIwa0Syi9g$>, and thank you for your pre-submission. No worries about inexperience - rOpenSci strives to be as welcoming and inclusive as possible, and to always be answer any questions you might have, and to help you along the way. As this is a pre-submission inquiry, feel free to ask questions here, or alternatively open issues in your own repository, cross-link them here by pasting the url for this issue in the comment, and ping me there. (Note that we try to keep the full submissions as "clean" as possible to help focus on reviews, while pre-submissions are the place for more general questions and dicussions.)

Specific responses to your questions:

  *   The "Error: No GNU global installation found" is because {pkgcheck}, and the {pkgstats} package it uses to analyses packages, require a couple of system libraries, including "GNU global". If you're on a Linux-based or MacOS, installation is simply, generally by using standard package manager (apt-get, homebrew, or whatever), to install global. If you're on Windows, it's tricker, but you could start with the links in the {pkgstats} installation vignette<https://urldefense.com/v3/__https:/docs.ropensci.org/pkgstats/articles/installation.html__;!!KGKeukY!1cCR1ET6HXFR0W8HksqghUebuvDfxPa2AmphB0DMkwS6B8XR4aTFPY3kwAa-MmzbsMJbCBpvou9AHHRYSIzKa2EYYA$>.
  *   The {srr} question sounds more general, and is maybe best moved to a specific issue within your repo? If you ping me there, I'll happily help further.

More generally, your package definitely looks like a good fit for statistical software review, and definitely within the category you've already indicated. Looking forward to working towards a full submission!
