### 2025-03-04 #690 @maurolepore
@mpadge what would you suggest is the next step forward? Our (amazing) dashboards highlights this as urgent because of inactivity, but maybe it already achieved its goal?
