### 2025-03-04 #690 @kelliejarcher
I used homebrew to install global and it is in my search path, as verified by

global --version

global (GNU Global) 6.6.14
Powered by Berkeley DB 1.85 and SQLite3 3.49.1.
Copyright (c) 1996-2024 Tama Communications Corporation
License GPLv3+: GNU GPL version 3 or later http://www.gnu.org/licenses/gpl.html
This is free software; you are free to change and redistribute it.
There is NO WARRANTY, to the extent permitted by law.

I also restarted my machine but am still getting
> x <- pkgcheck (mydir)
Error: No GNU global installation found.

I also have ctags installed:
ctags --version

Universal Ctags 6.1.0, Copyright (C) 2015-2023 Universal Ctags Team
Universal Ctags is derived from Exuberant Ctags.
Exuberant Ctags 5.8, Copyright (C) 1996-2009 Darren Hiebert
  Compiled: Feb 17 2025, 07:54:50
  URL: https://ctags.io/
  Output version: 0.0
  Optional compiled features: +wildcards, +regex, +gnulib_fnmatch, +gnulib_regex, +iconv, +option-directory, +xpath, +json, +interactive, +yaml, +case-insensitive-filenames, +packcc, +optscript, +pcre2

But in R when I run ctags_test() I get
Error: No GNU global installation found.

Do you know why this doesn’t work? My understanding is that I should not submit my package for ROpenSci review until I get this resolved - is that correct?
