### 2025-03-05 #690 @maurolepore
FYI I just added the holding label to test if this makes our software-review dashboard a bit happier :-)
I suspect the "no stage" is what weights the most here and disproportionately drives my attention.

![Image](https://github.com/user-attachments/assets/ca30196e-004c-4a1d-baea-10b23b14e3d6)
